+++ packages/components/src/components/modal/modal-styles.ts
@@ -99,6 +99,7 @@ export const getComponentCss = (
               placeSelf: 'stretch',
               margin: 0,
               borderRadius: 0,
+              clipPath: 'none', // fullscreen has square corners, so disable corner clipping
             }
           : {
               width: ref(cssVariableWidth, 'auto'),
@@ -107,6 +108,7 @@ export const getComponentCss = (
               placeSelf: 'center',
               margin: `${ref(cssVariableSpacingTop, 'clamp(16px, 10vh, 192px)')} ${gridExtendedOffsetBase} ${ref(cssVariableSpacingBottom, 'clamp(16px, 10vh, 192px)')}`, // horizontal margin is needed to ensure modal is placed on "Porsche Grid" when slotted content is wider than the viewport width
               borderRadius: dialogBorderRadius,
+              clipPath: `inset(0 round ${dialogBorderRadius})`, // non-fullscreen has rounded corners, so clip slotted content to them
               ...forcedColorsMediaQuery({
                 outline: '2px solid CanvasText',
                 outlineOffset: '-2px',